+++ script/test/fixtures/v3/docker-compose-env.yaml
@@ -1,0 +1,6 @@
+version: '3'
+services:
+  foo:
+    image: foo/bar:latest
+    environment:
+      FOO: foo
