+++ script/test/fixtures/v3/docker-compose-volumes.yaml
@@ -1,0 +1,8 @@
+version: "3"
+
+services:
+
+  foobar:
+    image: foo/bar:latest
+    volumes:
+      - /tmp/foo/bar
