+++ script/test/fixtures/v3/example2.env
@@ -1,0 +1,1 @@
+BAR=2
