+++ script/test/fixtures/v3/output-volumes-k8s.json
@@ -1,0 +1,106 @@
+{
+  "kind": "List",
+  "apiVersion": "v1",
+  "metadata": {},
+  "items": [
+    {
+      "kind": "Service",
+      "apiVersion": "v1",
+      "metadata": {
+        "name": "foobar",
+        "creationTimestamp": null,
+        "labels": {
+          "io.kompose.service": "foobar"
+        }
+      },
+      "spec": {
+        "ports": [
+          {
+            "name": "headless",
+            "port": 55555,
+            "targetPort": 0
+          }
+        ],
+        "selector": {
+          "io.kompose.service": "foobar"
+        },
+        "clusterIP": "None"
+      },
+      "status": {
+        "loadBalancer": {}
+      }
+    },
+    {
+      "kind": "Deployment",
+      "apiVersion": "extensions/v1beta1",
+      "metadata": {
+        "name": "foobar",
+        "creationTimestamp": null,
+        "labels": {
+          "io.kompose.service": "foobar"
+        }
+      },
+      "spec": {
+        "replicas": 1,
+        "template": {
+          "metadata": {
+            "creationTimestamp": null,
+            "labels": {
+              "io.kompose.service": "foobar"
+            }
+          },
+          "spec": {
+            "volumes": [
+              {
+                "name": "foobar-claim0",
+                "persistentVolumeClaim": {
+                  "claimName": "foobar-claim0"
+                }
+              }
+            ],
+            "containers": [
+              {
+                "name": "foobar",
+                "image": "foo/bar:latest",
+                "resources": {},
+                "volumeMounts": [
+                  {
+                    "name": "foobar-claim0",
+                    "mountPath": "/tmp/foo/bar"
+                  }
+                ]
+              }
+            ],
+            "restartPolicy": "Always"
+          }
+        },
+        "strategy": {
+          "type": "Recreate"
+        }
+      },
+      "status": {}
+    },
+    {
+      "kind": "PersistentVolumeClaim",
+      "apiVersion": "v1",
+      "metadata": {
+        "name": "foobar-claim0",
+        "creationTimestamp": null,
+        "labels": {
+          "io.kompose.service": "foobar-claim0"
+        }
+      },
+      "spec": {
+        "accessModes": [
+          "ReadWriteOnce"
+        ],
+        "resources": {
+          "requests": {
+            "storage": "100Mi"
+          }
+        }
+      },
+      "status": {}
+    }
+  ]
+}
